refactor: extract helpers out of component scope and remove dev code
- `validateEntry` and `isTimeRangeValid` moved out of BarberScheduleDialog
  body into module-level scope; `isTimeRangeValid` now lives in
  schedule-utils and is imported where needed
- `formatTimeLabel` extracted from `useSchedule` to module scope
- `getCopPrice` extracted from `ExtraCreditsCards` to module scope
- `oauthProviderLabel` extracted from `LoginPage` to module scope
- Removed `benchmarkDirections` dev utility from use-directions and its
  call site in BarbershopLocationSection
- Removed `process.env.NODE_ENV` guard from TanStack devtools in __root
- Fixed `message.tsx` AI bubble to use `secondary-foreground` in light
  mode so text is readable against the secondary background
- Import path for Convex types in barbershop-logo-uploader corrected to
  the `@convex` alias
